+++ tests/PowerUtils.Validations.Tests/Contracts/Strings/OptionsStringValidationsContractTests.cs
@@ -3,7 +3,7 @@ using PowerUtils.Validations;
 using PowerUtils.Validations.Tests.Fakes.Validations.Strings;
 using PowerUtils.Validations.Tests.Fakes.ValueObjects;
 
-namespace PowerUtils.RestAPI.Tests.Validations.Contracts.Strings;
+namespace PowerUtils.Validations.Tests.Contracts.Strings;
 
 public class OptionsStringValidationsContractTests
 {
