+++ tests/PowerUtils.Validations.Tests/Contracts/Strings/RequiredStringValidationsContractTests.cs
@@ -4,7 +4,7 @@ using PowerUtils.Validations.Contracts;
 using PowerUtils.Validations.Tests.Fakes.Entities;
 using PowerUtils.Validations.Tests.Fakes.Validations.Strings;
 
-namespace PowerUtils.RestAPI.Tests.Validations.Contracts.Strings;
+namespace PowerUtils.Validations.Tests.Contracts.Strings;
 
 public class RequiredStringValidationsContractTests
 {
